# Speed up JSON parsing