--- a/broadcast/broadcast.go
+++ b/broadcast/broadcast.go
@@ -10,6 +10,7 @@ import (
 	"github.com/segmentio/go-stats"
 	"github.com/segmentio/nsq_to_redis/ratelimit"
 	"github.com/statsd/client"
+	"github.com/tidwall/gjson"
 )
 
 type RedisPool interface {
@@ -24,8 +25,7 @@ type Handler interface {
 // Message is a parsed message.
 type Message struct {
 	ID   nsq.MessageID
-	Body []byte
-	JSON map[string]interface{}
+	JSON json.RawMessage
 }
 
 // Options for broadcast.
@@ -66,8 +66,7 @@ func (b *Broadcast) HandleMessage(msg *nsq.Message) error {
 	// parse
 	m := new(Message)
 	m.ID = msg.ID
-	m.Body = msg.Body
-	err := json.Unmarshal(m.Body, &m.JSON)
+	err := json.Unmarshal(msg.Body, &m.JSON)
 	if err != nil {
 		b.Log.Error("error parsing json: %s", err)
 		return nil
@@ -108,9 +107,8 @@ func (b *Broadcast) HandleMessage(msg *nsq.Message) error {
 // if ratelimit was not configured or exceeded.
 func (b *Broadcast) rateExceeded(msg *Message) bool {
 	if b.Ratelimiter != nil {
-		if v, ok := msg.JSON[b.RatelimitKey].(string); ok {
-			return b.Ratelimiter.Exceeded(v)
-		}
+		k := gjson.Get(string(msg.JSON), b.RatelimitKey).String()
+		return b.Ratelimiter.Exceeded(k)
 	}
 
 	return false
@@ -119,18 +117,15 @@ func (b *Broadcast) rateExceeded(msg *Message) bool {
 // NewMessage returns a Message or an error if unable to do so.
 // Used primarily by tests.
 func NewMessage(id, contents string) (*Message, error) {
-	var body map[string]interface{}
-	err := json.Unmarshal([]byte(contents), &body)
+	nsqId := [nsq.MsgIDLength]byte{}
+	copy(nsqId[:], id[:nsq.MsgIDLength])
+
+	m := new(Message)
+	m.ID = nsqId
+	err := json.Unmarshal([]byte(contents), &m.JSON)
 	if err != nil {
 		return nil, err
 	}
 
-	nsqId := [nsq.MsgIDLength]byte{}
-	copy(nsqId[:], id[:nsq.MsgIDLength])
-
-	return &Message{
-		ID:   nsqId,
-		Body: []byte(contents),
-		JSON: body,
-	}, nil
+	return m, nil
 }