--- a/list/list.go
+++ b/list/list.go
@@ -3,10 +3,10 @@ package list
 import (
 	"time"
 
-	interpolate "github.com/segmentio/go-interpolate"
 	"github.com/segmentio/go-log"
 	"github.com/segmentio/go-stats"
 	"github.com/segmentio/nsq_to_redis/broadcast"
+	"github.com/segmentio/nsq_to_redis/template"
 	"github.com/statsd/client"
 )
 
@@ -20,7 +20,7 @@ type Options struct {
 
 // List writes messages to capped lists.
 type List struct {
-	template *interpolate.Template
+	template *template.T
 	stats    *stats.Stats
 	*Options
 }
@@ -32,7 +32,7 @@ func New(options *Options) (*List, error) {
 		stats:   stats.New(),
 	}
 
-	tmpl, err := interpolate.New(r.Format)
+	tmpl, err := template.New(r.Format)
 	if err != nil {
 		return nil, err
 	}
@@ -49,16 +49,16 @@ func New(options *Options) (*List, error) {
 func (l *List) Handle(c broadcast.Conn, msg *broadcast.Message) error {
 	start := time.Now()
 
-	key, err := l.template.Eval(msg.JSON)
+	key, err := l.template.Eval(string(msg.JSON))
 	if err != nil {
 		l.Log.Error("evaluating template: %s", err)
 		return nil
 	}
 
 	l.Log.Info("pushing %s to %s", msg.ID, key)
-	l.Log.Debug("contents %s %s", msg.ID, msg.Body)
+	l.Log.Debug("contents %s %s", msg.ID, msg.JSON)
 
-	err = c.Send("LPUSH", key, []byte(msg.Body))
+	err = c.Send("LPUSH", key, []byte(msg.JSON))
 	if err != nil {
 		l.Log.Error("lpush: %s", err)
 	}