This is basically #11 but re-introduced now that we have better benchmarks
and tests.

```
 $ benchcmp old-interpolate.txt new-gjson.txt
 benchmark                old ns/op     new ns/op     delta
 Benchmark1handler-8      2526          1800          -28.74%
 Benchmark2Handlers-8     2670          1813          -32.10%
 BenchmarkList-8          68003         65095         -4.28%
 BenchmarkPubSub-8        32610         32462         -0.45%
```